Serialization tests

+++ inc/messages/frame.h
@@ -6,11 +6,11 @@
 class frame: public serializable {
 	public:
 		uint32_t time;
-		uint8_t tag;
+		uint16_t tag;
 		serializable* payload;
 		
 		frame();
-		frame(uint32_t time, uint8_t tag, serializable* payload);
+		frame(uint32_t time, uint16_t tag, serializable* payload);
 		virtual void serialize(uint8_t* buf);
 		virtual void deserialize(uint8_t* serialized);
 		virtual size_t get_size();
